이인아 Q&A 게시판 CRUD 기능 구현 완료, 운동페이지 캘린더 이벤트 기능 구현 시작

+++ src/main/java/com/aws/carepoint/dto/QnaDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class QnaDto {
-    private Long articlePk;
+    private int articlePk;
     private String content;
     private String filename;
     private Integer recom;
@@ -16,7 +16,7 @@ public class QnaDto {
     private LocalDateTime regDate;
     private LocalDateTime updateDate;
     private String userNick;
-    private Long boardPk;
+    private int boardPk;
     private String title;
     private String boardType;
     private Integer userPk;
